Split parse/ast/nodes module into nodes/ package with facade
